feat: move components to specific subfolder
